Refactor to move some stuff to utils
